pi/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
websocket-client>=1.6.0
wakeonlan>=3.1.0
joblib>=1.3.0
